+++ python/dartpy/dynamics/Frame.cpp
@@ -36,6 +36,7 @@
 #include <dart/dart.hpp>
 
 #include <pybind11/pybind11.h>
+#include <pybind11/stl.h>
 
 namespace py = pybind11;